refactor hunter
